feat: refatoração
